--- packages/functions/.env.example
@@ -1,1 +1,0 @@
-INFURA_PROJECT_ID=
