--- packages/functions/.eslintrc.json
@@ -1,3 +1,0 @@
-{
-  "extends": ["../../.eslintrc.json"]
-}
