--- packages/functions/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@kaguraswap/functions",
-  "version": "0.0.1",
-  "license": "MIT",
-  "private": true,
-  "scripts": {
-    "dev": "tsc --watch",
-    "build": "tsc"
-  },
-  "main": "dist/functions/src/index.js",
-  "dependencies": {
-    "@opensea/seaport-js": "^1.0.2",
-    "axios": "^0.27.2",
-    "canonicalize": "^1.0.8",
-    "cors": "^2.8.5",
-    "ethers": "^5.6.9",
-    "firebase-admin": "^10.0.2",
-    "firebase-functions": "^3.18.0",
-    "ipfs-only-hash": "^4.0.0"
-  },
-  "devDependencies": {
-    "firebase-functions-test": "^0.2.0",
-    "typescript": "^4.5.4"
-  }
-}
