--- packages/functions/src/lib/cors.ts
@@ -1,3 +1,0 @@
-import _cors from "cors";
-
-export const cors = _cors({ origin: true });
