--- packages/functions/src/lib/functions.ts
@@ -1,5 +1,0 @@
-import * as _functions from "firebase-functions";
-
-import { FIREBASE_REGION } from "../../../common/configs/app";
-
-export const functions = _functions.region(FIREBASE_REGION);
