--- packages/functions/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "sourceMap": true
-  },
-  "compileOnSave": true,
-  "include": ["src"]
-}
