+++ packages/web/package.json
@@ -29,12 +29,14 @@
     "framer-motion": "^6",
     "graphql": "^16.5.0",
     "next": "12.2.0",
+    "pg": "^8.7.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-firebase-hooks": "^5.0.3",
     "react-icons": "^4.4.0",
     "react-share": "^4.4.0",
-    "react-spinners": "^0.13.3"
+    "react-spinners": "^0.13.3",
+    "sequelize": "^6.21.3"
   },
   "devDependencies": {
     "@babel/core": "^7.18.6",
